Changed the formatline function argument type for the Attr because of the change in SET's editor

+++ ideforma.cc
@@ -38,7 +38,7 @@ int DebuggerFormatLine(TCEditor *editor,
                        int Width,
                        unsigned short Colors,
                        unsigned lineLen,
-                       unsigned short Attr,
+                       uint32 Attr,
                        unsigned LineNo)
 {
 #define drawbuf ((ushort *)DrawBuf)
